Improve performance in several key areas
After running some profiling I identifier the pain points of
the matcher. One of the solutions (binary search) seems
to speed it up in a major way, while the other one (splitting
on unicodescalars instead of normal characters) made a big
difference in profiling but not in actual performance testing.
Not sure what to think of this but either way the performance
of the matcher is about 2-3 times faster.